embedded_legacy_form.ftl: removing comment; exporting guestMessages.ts to assets; adding DialogProps to BrowseFilesDialog.tsx; creating BrowseFilesDialogUI.tsx; splitting out FolderBrowserTreeView.tsx
