+++ ui/app/src/components/BrowseFilesDialog/index.ts
@@ -17,3 +17,5 @@
 export { default } from './BrowseFilesDialog';
 
 export * from './BrowseFilesDialogContainer';
+
+export * from './BrowseFilesDialogUI';
