+++ ui/app/src/components/Dialogs/PathSelectionDialog.tsx
@@ -20,7 +20,7 @@ import DialogFooter from './DialogFooter';
 import React, { PropsWithChildren, useEffect, useMemo, useRef, useState } from 'react';
 import { FormattedMessage } from 'react-intl';
 import Dialog from '@material-ui/core/Dialog';
-import FolderBrowserTreeViewUI, { legacyItemsToTreeNodes, TreeNode } from '../Navigation/FolderBrowserTreeViewUI';
+import FolderBrowserTreeViewUI, { TreeNode } from '../FolderBrowserTreeView/FolderBrowserTreeViewUI';
 import { createStyles, makeStyles } from '@material-ui/core/styles';
 import CreateFolderDialog from './CreateFolderDialog';
 import { get } from '../../utils/ajax';
@@ -38,6 +38,7 @@ import { useActiveSiteId } from '../../utils/hooks/useActiveSiteId';
 import { useLogicResource } from '../../utils/hooks/useLogicResource';
 import { useUnmount } from '../../utils/hooks/useUnmount';
 import { usePossibleTranslation } from '../../utils/hooks/usePossibleTranslation';
+import { legacyItemsToTreeNodes } from '../FolderBrowserTreeView/utils';
 
 export interface PathSelectionDialogBaseProps {
   open: boolean;
