+++ ui/app/src/modules/Preview/PreviewConcierge.tsx
@@ -73,7 +73,7 @@ import { filter, map, pluck, switchMap, take, takeUntil } from 'rxjs/operators';
 import ContentType from '../../models/ContentType';
 import { forkJoin, Observable, of, ReplaySubject } from 'rxjs';
 import Button from '@material-ui/core/Button';
-import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
+import { FormattedMessage, useIntl } from 'react-intl';
 import { getGuestToHostBus, getHostToGuestBus, getHostToHostBus } from './previewContext';
 import { useDispatch } from 'react-redux';
 import { findParentModelId, nnou, pluckProps } from '../../utils/object';
@@ -116,89 +116,7 @@ import { showEditDialog } from '../../state/actions/dialogs';
 import { UNDEFINED } from '../../utils/constants';
 import { useCurrentPreviewItem } from '../../utils/hooks/useCurrentPreviewItem';
 import { useSiteUIConfig } from '../../utils/hooks/useSiteUIConfig';
-
-export const guestMessages = defineMessages({
-  maxCount: {
-    id: 'validations.maxCount',
-    defaultMessage: 'The max number of items is {maxCount}'
-  },
-  minCount: {
-    id: 'validations.minCount',
-    defaultMessage: 'The min number of items is {minCount}'
-  },
-  required: {
-    id: 'validations.required',
-    defaultMessage: '{field} is required'
-  },
-  maxLength: {
-    id: 'validations.maxLength',
-    defaultMessage: 'The max length ({maxLength}) reached'
-  },
-  yes: {
-    id: 'words.yes',
-    defaultMessage: 'Yes'
-  },
-  registerNotFound: {
-    id: 'register.notFound',
-    defaultMessage: '{name} is not visible or was not registered by developers'
-  },
-  dropTargetsNotFound: {
-    id: 'register.dropTargetsNotFound',
-    defaultMessage: 'There are no drop targets for {contentType} components'
-  },
-  sortOperationComplete: {
-    id: 'operations.sortOperationComplete',
-    defaultMessage: 'Sort operation completed.'
-  },
-  sortOperationFailed: {
-    id: 'operations.sortOperationFailed',
-    defaultMessage: 'Sort operation failed.'
-  },
-  insertOperationComplete: {
-    id: 'operations.insertOperationComplete',
-    defaultMessage: 'Insert component operation completed.'
-  },
-  insertOperationFailed: {
-    id: 'operations.insertOperationFailed',
-    defaultMessage: 'Insert component operation failed.'
-  },
-  insertItemOperation: {
-    id: 'operations.insertItemOperation',
-    defaultMessage: 'Insert item operation not implemented.'
-  },
-  moveOperationComplete: {
-    id: 'operations.moveOperationComplete',
-    defaultMessage: 'Move operation completed'
-  },
-  moveOperationFailed: {
-    id: 'operations.moveOperationFailed',
-    defaultMessage: 'Move operation failed.'
-  },
-  deleteOperationComplete: {
-    id: 'operations.deleteOperationComplete',
-    defaultMessage: 'Delete operation completed.'
-  },
-  deleteOperationFailed: {
-    id: 'operations.deleteOperationFailed',
-    defaultMessage: 'Delete operation failed.'
-  },
-  updateOperationComplete: {
-    id: 'operations.updateOperationComplete',
-    defaultMessage: 'Update operation completed.'
-  },
-  updateOperationFailed: {
-    id: 'operations.updateOperationFailed',
-    defaultMessage: 'Update operation failed.'
-  },
-  assetUploadStarted: {
-    id: 'operations.assetUploadStarted',
-    defaultMessage: 'Asset upload started.'
-  },
-  assetUploadFailed: {
-    id: 'operations.assetUploadFailed',
-    defaultMessage: 'Asset Upload failed.'
-  }
-});
+import { guestMessages } from '../../assets/guestMessages';
 
 const originalDocDomain = document.domain;
